+++ src/types/trash.d.ts
@@ -1,0 +1,7 @@
+declare module 'trash' {
+  type TrashOptions = {
+    glob?: boolean;
+  };
+
+  export default function trash(paths: string[], options?: TrashOptions): Promise<string[]>;
+}